Se arreglo desplegables y la actualizacion en la base de datos y se cambiaron sql que no andaban bien se relaciono y se completo el articulado agregado con id respectivos

+++ Dominio/Articulo.cs
@@ -15,7 +15,7 @@ namespace Dominio
         public string descripcion { get; set; } 
         public Marca marca { get; set; } //(seleccionable de una lista desplegable).
         public Categoria categoria { get; set; } //(seleccionable de una lista desplegable.
-        public float precio { get; set; }
+        public decimal precio { get; set; }
         public string URLimagen {  get; set; }
   
     }
